+++ eras/conway/impl/src/Cardano/Ledger/Conway/Scripts.hs
@@ -53,7 +53,7 @@ import Cardano.Ledger.Plutus.Language
 import Cardano.Ledger.Shelley.Scripts (ShelleyEraScript (..))
 import Cardano.Ledger.TxIn (TxIn)
 import Control.DeepSeq (NFData (..), rwhnf)
-import Data.Aeson (ToJSON (..), (.=))
+import Data.Aeson (FromJSON (..), ToJSON (..), withObject, (.:), (.=))
 import Data.MemPack
 import Data.Typeable
 import Data.Word (Word32)
@@ -311,6 +311,25 @@ instance
     ConwayProposing n -> kindObjectWithValue "ConwayProposing" n
     where
       kindObjectWithValue name n = kindObjectValue name ["value" .= n]
+
+instance
+  ( forall a b. (FromJSON a, FromJSON b) => FromJSON (f a b)
+  , FromJSON (TxCert era)
+  , EraPParams era
+  ) =>
+  FromJSON (ConwayPlutusPurpose f era)
+  where
+  parseJSON = withObject "ConwayPlutusPurpose" $ \o -> do
+    kind <- o .: "kind"
+    value <- o .: "value"
+    case (kind :: String) of
+      "ConwaySpending" -> ConwaySpending <$> parseJSON value
+      "ConwayMinting" -> ConwayMinting <$> parseJSON value
+      "ConwayCertifying" -> ConwayCertifying <$> parseJSON value
+      "ConwayRewarding" -> ConwayRewarding <$> parseJSON value
+      "ConwayVoting" -> ConwayVoting <$> parseJSON value
+      "ConwayProposing" -> ConwayProposing <$> parseJSON value
+      _ -> fail $ "Unknown ConwayPlutusPurpose kind: " <> kind
 
 pattern VotingPurpose ::
   ConwayEraScript era => f Word32 Voter -> PlutusPurpose f era